# perf: benchmark v2 pipeline Backends.ingest_logs/2