```
Name                                                  ips        average  deviation         median         99th %
SourceSup v2 BQ with Backends.ingest_logs/2       1264.06        0.79 ms    ±34.97%        0.80 ms        1.68 ms
SourceSup v2 BQ with Logs.ingest_logs/2            370.75        2.70 ms    ±61.23%        2.63 ms        5.59 ms
v1SourceSup BQ with Logs.ingest_logs/2             369.72        2.70 ms    ±61.33%        2.67 ms        5.83 ms

Comparison: 
SourceSup v2 BQ with Backends.ingest_logs/2       1264.06
SourceSup v2 BQ with Logs.ingest_logs/2            370.75 - 3.41x slower +1.91 ms
v1SourceSup BQ with Logs.ingest_logs/2             369.72 - 3.42x slower +1.91 ms

Extended statistics: 

Name                                                minimum        maximum    sample size                     mode
SourceSup v2 BQ with Backends.ingest_logs/2         0.29 ms        3.00 ms            518 0.47 ms, 0.92 ms, 0.56 ms
SourceSup v2 BQ with Logs.ingest_logs/2             0.21 ms       11.37 ms            417 2.48 ms, 2.63 ms, 0.98 ms
v1SourceSup BQ with Logs.ingest_logs/2             0.120 ms       14.14 ms            425                     None
```

Backends.ingest_logs/2 is 3.4x faster when handling event batches (n=150)

BQ insert rate is also at 1.2x higher throughput (518/425) assuming perfect conditions.

V2 should give good perf gains when handling raw requests. 🎉